backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy[asyncio]==2.0.23
aiosqlite==0.19.0
openai==1.3.7
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6
python-jose[cryptography]==3.5.0
bcrypt==4.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pydantic==2.5.0
asyncpg>=0.28.0
alembic